CinematicSound: Use size_t

+++ src/cinematic/CinematicSound.cpp
@@ -84,7 +84,7 @@ static CinematicSound * GetFreeSound() {
 	return NULL;
 }
 
-static bool DeleteFreeSound(int num) {
+static bool DeleteFreeSound(size_t num) {
 	
 	if(!TabSound[num].exists) {
 		return false;
